feat: Complete Phase 3 - Move final components
